+++ src/compiler/link.h
@@ -26,4 +26,6 @@ bool           LI_segmentListReadObjectFile (LI_segmentList sl, string sourcefn,
 bool           LI_link                      (LI_segmentList sl);
 void           LI_segmentListWriteLoadFile  (LI_segmentList sl, string loadfn);
 
+bool           LI_segmentListReadLoadFile   (LI_segmentList sl, string sourcefn, FILE *f);
+
 #endif