+++ CHANGELOG.md
@@ -7,7 +7,7 @@ Improvements:
     * ide: exit or continue choice on traps and debug breaks
     * ide: SMART\_REFRESH window for faster+reliable refresh handling
     * compiler: OPTION DEBUG statement added
-    * compiler: DPRINT statement added
+    * compiler: TRACE statement added
     * compiler: BREAK statement added
     * runtime: CTRL-C will issue a DEBUG BREAK when debugger is active
     * runtime: BITMAP\_t supports continous plane allocation suitable for BOB use now